requirements.txt
tqdm
pillow
numpy
pandas
geopandas
rasterio
requests
pydap
xarray
cdsapi
scikit-learn
uptide
webob
netCDF4